// src/state/types.ts
export type SerializedBigNumber = string

interface SerializedFarmUserData {
    allowance: string
    tokenBalance: string
    stakedBalance: string
    earnings: string
}

export interface DeserializedFarmUserData {
    allowance: BigNumber
    tokenBalance: BigNumber
    stakedBalance: BigNumber
    earnings: BigNumber
}

export interface SerializedFarm extends SerializedFarmConfig {
    tokenPriceBusd?: string
    quoteTokenPriceBusd?: string
    tokenAmountTotal?: SerializedBigNumber
    lpTotalInQuoteToken?: SerializedBigNumber
    lpTotalSupply?: SerializedBigNumber
    tokenPriceVsQuote?: SerializedBigNumber
    poolWeight?: SerializedBigNumber
    userData?: SerializedFarmUserData
}

export interface Game {
    title: string
    subtitle: string
    logo: string
    cta: string
    symbol: string
    votes: number
    chain: string
    category?: GameCategories
    address?: string
    twitter?: string
    telegram?: string
    discord?: string
    platform?: GamePlatforms[]
}

export interface DeserializedFarm extends DeserializedFarmConfig {
    tokenPriceBusd?: string
    quoteTokenPriceBusd?: string
    tokenAmountTotal?: BigNumber
    lpTotalInQuoteToken?: BigNumber
    lpTotalSupply?: BigNumber
    tokenPriceVsQuote?: BigNumber
    poolWeight?: BigNumber
    userData?: DeserializedFarmUserData
}
// ...
export interface SerializedPool extends SerializedPoolConfig, CorePoolProps {
    totalStaked?: SerializedBigNumber
    stakingLimit?: SerializedBigNumber
    userData?: {
        allowance: SerializedBigNumber
        stakingTokenBalance: SerializedBigNumber
        stakedBalance: SerializedBigNumber
        pendingReward: SerializedBigNumber
    }
}

export interface Profile {
    userId: number
    points: number
    teamId: number
    collectionAddress: string
    tokenId: number
    isActive: boolean
    username: string
    nft?: NftToken
    team: Team
    hasRegistered: boolean
}

// Slices states

export interface SerializedFarmsState {
    data: SerializedFarm[]
    loadArchivedFarmsData: boolean
    userDataLoaded: boolean
    loadingKeys: Record<string, boolean>
}

export interface GamesState {
    data: Game[]
    userDataLoaded: boolean
    isLoading: boolean
    prices: {
        address: string
        price: {
            bnb: number
            usd: number
        }
    }[]
}

export interface DeserializedFarmsState {
    data: DeserializedFarm[]
    loadArchivedFarmsData: boolean
    userDataLoaded: boolean
}

export interface VaultFees {
    performanceFee: number
    callFee: number
    withdrawalFee: number
    withdrawalFeePeriod: number
}

export interface VaultUser {
    isLoading: boolean
    userShares: string
    cakeAtLastUserAction: string
    lastDepositedTime: string
    lastUserActionTime: string
}

export interface IfoVaultUser extends VaultUser {
    credit: string
}

export interface CakeVault {
    totalShares?: string
    pricePerFullShare?: string
    totalCakeInVault?: string
    estimatedCakeBountyReward?: string
    totalPendingCakeHarvest?: string
    fees?: VaultFees
    userData?: VaultUser
}

export interface IfoCakeVault extends Omit<CakeVault, 'userData'> {
    userData?: IfoVaultUser
    creditStartBlock?: number
}

export interface PoolsState {
    data: SerializedPool[]
    cakeVault: CakeVault
    ifoPool: IfoCakeVault
    userDataLoaded: boolean
}

export enum ProfileAvatarFetchStatus {
    NOT_FETCHED = 'not-fetched',
    FETCHING = 'fetching',
    FETCHED = 'fetched',
}

export interface ProfileState {
    isInitialized: boolean
    isLoading: boolean
    hasRegistered: boolean
    data: Profile | null
    profileAvatars: {
        [key: string]: {
            username: string | null
            nft: NftToken | null
            hasRegistered: boolean
            usernameFetchStatus: ProfileAvatarFetchStatus
            avatarFetchStatus: ProfileAvatarFetchStatus
        }
    }
}

export type TeamResponse = {
    0: string
    1: string
    2: string
    3: string
    4: boolean
}

export type TeamsById = {
    [key: string]: Team
}

export interface TeamsState {
    isInitialized: boolean
    isLoading: boolean
    data: TeamsById
}

export interface Achievement {
    id: string
    type: CampaignType
    address: string
    title: TranslatableText
    description?: TranslatableText
    badge: string
    points: number
}

export enum AchievementFetchStatus {
    ERROR = 'error',
    NOT_FETCHED = 'not-fetched',
    FETCHING = 'fetching',
    FETCHED = 'fetched',
}

export interface AchievementState {
    achievements: Achievement[]
    achievementFetchStatus: AchievementFetchStatus
}

// Block

export interface BlockState {
    currentBlock: number
    initialBlock: number
}

// Predictions

export enum BetPosition {
    BULL = 'Bull',
    BEAR = 'Bear',
    HOUSE = 'House',
}

export enum PredictionStatus {
    INITIAL = 'initial',
    LIVE = 'live',
    PAUSED = 'paused',
    ERROR = 'error',
}
// ...
export interface Bet {
    id?: string
    hash?: string
    amount: number
    position: BetPosition
    claimed: boolean
    claimedAt: number
    claimedBlock: number
    claimedHash: string
    claimedBNB: number
    claimedNetBNB: number
    createdAt: number
    updatedAt: number
    user?: PredictionUser
    round?: Round
}

export interface PredictionUser {
    id: string
    createdAt: number
    updatedAt: number
    block: number
    totalBets: number
    totalBetsBull: number
    totalBetsBear: number
    totalBNB: number
    totalBNBBull: number
    totalBNBBear: number
    totalBetsClaimed: number
    totalBNBClaimed: number
    winRate: number
    averageBNB: number
    netBNB: number
    bets?: Bet[]
}

export enum HistoryFilter {
    ALL = 'all',
    COLLECTED = 'collected',
    UNCOLLECTED = 'uncollected',
}

export interface LedgerData {
    [key: string]: {
        [key: string]: ReduxNodeLedger
    }
}

export interface RoundData {
    [key: string]: ReduxNodeRound
}
// ...
export enum LeaderboardLoadingState {
    INITIAL,
    LOADING,
    IDLE,
}

export type LeaderboardFilterTimePeriod = '1d' | '7d' | '1m' | 'all'

export interface LeaderboardFilter {
    address?: string
    orderBy?: string
    timePeriod?: LeaderboardFilterTimePeriod
}

export interface PredictionsState {
    status: PredictionStatus
    isLoading: boolean
    isHistoryPaneOpen: boolean
    isChartPaneOpen: boolean
    isFetchingHistory: boolean
    historyFilter: HistoryFilter
    currentEpoch: number
    intervalSeconds: number
    minBetAmount: string
    bufferSeconds: number
    lastOraclePrice: string
    history: Bet[]
    totalHistory: number
    currentHistoryPage: number
    hasHistoryLoaded: boolean
    rounds?: RoundData
    ledgers?: LedgerData
    claimableStatuses: {
        [key: string]: boolean
    }
    leaderboard: {
        selectedAddress: string
        loadingState: LeaderboardLoadingState
        filters: LeaderboardFilter
        skip: number
        hasMoreResults: boolean
        addressResults: {
            [key: string]: PredictionUser
        }
        results: PredictionUser[]
    }
}
// ...
export interface Proposal {
    author: string
    body: string
    choices: string[]
    end: number
    id: string
    snapshot: string
    space: Space
    start: number
    state: ProposalState
    title: string
}

export interface Vote {
    id: string
    voter: string
    created: number
    space: Space
    proposal: {
        choices: Proposal['choices']
    }
    choice: number
    metadata?: {
        votingPower: string
        verificationHash: string
    }
    _inValid?: boolean
}

export enum VotingStateLoadingStatus {
    INITIAL = 'initial',
    IDLE = 'idle',
    LOADING = 'loading',
    ERROR = 'error',
}

export interface VotingState {
    proposalLoadingStatus: VotingStateLoadingStatus
    proposals: {
        [key: string]: Proposal
    }
    voteLoadingStatus: VotingStateLoadingStatus
    votes: {
        [key: string]: Vote[]
    }
}

export interface LotteryRoundUserTickets {
    isLoading?: boolean
    tickets?: LotteryTicket[]
}
// ...
export interface LotteryResponse extends LotteryRoundGenerics {
    priceTicketInCake: SerializedBigNumber
    discountDivisor: SerializedBigNumber
    amountCollectedInCake: SerializedBigNumber
    cakePerBracket: SerializedBigNumber[]
    countWinnersPerBracket: SerializedBigNumber[]
    rewardsBreakdown: SerializedBigNumber[]
}

export interface LotteryState {
    currentLotteryId: string
    maxNumberTicketsPerBuyOrClaim: string
    isTransitioning: boolean
    currentRound: LotteryResponse & { userTickets?: LotteryRoundUserTickets }
    lotteriesData?: LotteryRoundGraphEntity[]
    userLotteryData?: LotteryUserGraphEntity
}

export interface LotteryRoundGraphEntity {
    id: string
    totalUsers: string
    totalTickets: string
    winningTickets: string
    status: LotteryStatus
    finalNumber: string
    startTime: string
    endTime: string
    ticketPrice: SerializedBigNumber
}

export interface LotteryUserGraphEntity {
    account: string
    totalCake: string
    totalTickets: string
    rounds: UserRound[]
}

export interface UserRound {
    claimed: boolean
    lotteryId: string
    status: LotteryStatus
    endTime: string
    totalTickets: string
    tickets?: LotteryTicket[]
}

export type UserTicketsResponse = [ethers.BigNumber[], number[], boolean[]]

// Global state

export interface State {
    achievements: AchievementState
    block: BlockState
    farms: SerializedFarmsState
    games: GamesState
    pools: PoolsState
    predictions: PredictionsState
    profile: ProfileState
    teams: TeamsState
    voting: VotingState
    lottery: LotteryState
    nftMarket: NftMarketState
}